chore: Add permission for ci (#562)
## What does this change?

Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

Fixes #<issue_number_goes_here> 🎯

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

## Checklist:

Before submitting your PR, there are a few things you can do to make
sure it goes smoothly:

- [ ] Make sure to open a GitHub issue as a bug/feature request before
writing your code! That way we can discuss the change, evaluate designs,
and agree on the general idea
  - [ ] Ensure the tests (`melos run test`)
- [ ] Ensure the analyzer and formatter pass (`melos run format` to
automatically apply formatting)
- [ ] Appropriate docs were updated (if necessary)