⚡️ Enables `gaplessPlayback` by default for image assets (#569)
## What does this change?

Image assets are finalized during builds, they cannot be changed during
the runtime. The `gaplessPlayback` defaults to `false` which might lead
to images flashing between routes when the provider changes, which it
doesn't in fact.

Now we change it to `true` to ignore changes to provide a good display
experience. This can also considered as a non-breaking change in most
cases.

## Type of change

- [x] New feature (non-breaking change which adds functionality)