Add support for deferred components (#576)
## What does this change?

This PR include support for [Flutter Deferred Components](https://docs.flutter.dev/perf/deferred-components).

Deferred components allow developers to the app into multiple apk to reduce its size. This can be used to optimize the initial download and download components only on necessary but it is also mandatory for apk with more than 200MB.

This library initially looks for assets listed on `flutter.assets` of `pubspec.yaml`, with this change the library will continue looking for the assets lists under `flutter.assets` but will merge with the list of assets included on `flutter.deferred-components.$`. This will be merged in a single list.

Ideally in the future assets should have different classname, so users will easily know when dealing with deferred components. This wasn't implemented at this moment because it would requeire a significative changes.

Fixes #577 🎯

## Type of change

- [x] New feature (non-breaking change which adds functionality)
- [x] This change requires a documentation update

---------

Co-authored-by: Ian Maciel <ian.maciel@motorolasolutions.com>
Co-authored-by: Alex Li <github@alexv525.com>